+++ app/controllers/journeys_controller.rb
@@ -5,19 +5,31 @@ class JourneysController < ApplicationController
     render "errors/specification_template_invalid", status: 500, locals: {error: exception}
   end
 
+  rescue_from CreateJourneyStep::UnexpectedContentfulModel do |exception|
+    render "errors/unexpected_contentful_model", status: 500
+  end
+
+  rescue_from CreateJourneyStep::UnexpectedContentfulStepType do |exception|
+    render "errors/unexpected_contentful_step_type", status: 500
+  end
+
+  rescue_from BuildJourneyOrder::MissingEntryDetected do |exception|
+    render "errors/contentful_entry_not_found", status: 500
+  end
+
   def new
     journey = CreateJourney.new(category: "catering").call
-    redirect_to new_journey_step_path(journey)
+    redirect_to journey_path(journey)
   end
 
   def show
     @journey = Journey.includes(
-      steps: [:radio_answer, :short_text_answer, :long_text_answer]
+      steps: [:radio_answer, :short_text_answer, :long_text_answer, :single_date_answer, :checkbox_answers]
     ).find(journey_id)
     @steps = @journey.steps.map { |step| StepPresenter.new(step) }
 
     @answers = @journey.steps.that_are_questions.each_with_object({}) { |step, hash|
-      hash["answer_#{step.contentful_id}"] = step.answer.response.to_s
+      hash["answer_#{step.contentful_id}"] = step.answer&.response.to_s
     }
 
     @specification_template = Liquid::Template.parse(