+++ app/controllers/steps_controller.rb
@@ -1,18 +1,6 @@
 # frozen_string_literal: true
 
 class StepsController < ApplicationController
-  rescue_from GetContentfulEntry::EntryNotFound do |exception|
-    render "errors/contentful_entry_not_found", status: 500
-  end
-
-  rescue_from CreateJourneyStep::UnexpectedContentfulModel do |exception|
-    render "errors/unexpected_contentful_model", status: 500
-  end
-
-  rescue_from CreateJourneyStep::UnexpectedContentfulStepType do |exception|
-    render "errors/unexpected_contentful_step_type", status: 500
-  end
-
   def new
     @journey = Journey.find(journey_id)